refactor(model): replace per-cell CellBC with per-face BoundaryPatch
The old design allocated 6 (CellBC::types + CellBC::param_idxs) per active
cell — O(N) storage when only O(N^{2/3}) faces are on domain boundaries.

Key changes:
- Remove struct CellBC from model.hpp
- Add struct BoundaryPatch {cell_idx, dir, type, param_idx}
- Model::boundary_patches — flat list of all boundary faces
- CellFields::cell_bc_range — prefix-sum index into boundary_patches
- Keep BCParamTable (unchanged) for BC expression storage

Preprocessor (layer_processor):
- apply_face_bcs_to_cell → apply_face_bc_to_patch, pushes BoundaryPatch
- resolve_layers now takes boundary_patches out-parameter
- Three-phase: count exposed faces → build cell_bc_range → fill patches

Assembler (assembler.cpp):
- Split into two TBB parallel_for phases:
  Phase 1: internal conduction + heat source + mass + fluid advection
            (skip exposed faces — no BC query needed during cell sweep)
  Phase 2: iterate model_.boundary_patches, accumulate BC terms

Postprocessor / probe_recorder:
- Query patches via cell_bc_range prefix-sum instead of cell_bcs[]

All 122 tests pass; all steady & transient cases pass.

+++ src/data/types.hpp
@@ -13,13 +13,13 @@ namespace mhs::core {
     // Fluid boundary condition type. Independent of BcType — the same face
     // can carry both a thermal BC and a fluid BC simultaneously.
     enum class FluidBCType : uint8_t {
-        None             = 0,
-        PressureType     = 1, // Dirichlet on p
+        None = 0,
+        PressureType = 1, // Dirichlet on p
         MassFlowRateType = 2, // Neumann: m_dot [kg/s] -> overrides energy netOutflux
-        VelocityType     = 3, // Neumann: u [m/s] normal -> overrides energy netOutflux
+        VelocityType = 3, // Neumann: u [m/s] normal -> overrides energy netOutflux
     };
 
-    enum class FaceDir : size_t { XM = 0, XP = 1, YM = 2, YP = 3, ZM = 4, ZP = 5 };
+    enum class FaceDir : uint8_t { XM = 0, XP = 1, YM = 2, YP = 3, ZM = 4, ZP = 5 };
 
     constexpr size_t FACE_COUNT = 6;
